Completed unification of systems and other locations, and into single "discover" page flow
